requirements.txt
yt-dlp
streamlit
pytube
requests
pillow